found the bug with saving a new note

+++ LocalNote/Commands/DeleteCommand.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Input;
+using Windows.UI.Xaml.Controls;
 
 namespace LocalNote.Commands
 {
@@ -11,6 +12,7 @@ namespace LocalNote.Commands
     {
         public event EventHandler CanExecuteChanged;
         private readonly ViewModels.NoteViewModel noteViewModel;
+        private readonly Views.DeleteNoteDialog delete;
 
         /// <summary>
         /// Constructor
@@ -19,6 +21,7 @@ namespace LocalNote.Commands
         public DeleteCommand(ViewModels.NoteViewModel noteViewModel)
         {
             this.noteViewModel = noteViewModel;
+            this.delete = new Views.DeleteNoteDialog();
         }
 
         /// <summary>
@@ -41,30 +44,35 @@ namespace LocalNote.Commands
         /// Executes the command.
         /// </summary>
         /// <param name="parameter"></param>
-        public void Execute(object parameter)
+        public async void Execute(object parameter)
         {
-            // Get the selected note
-            Models.NoteModel noteToDelete = this.noteViewModel.SelectedNote;
+            ContentDialogResult result = await delete.ShowAsync();
 
-            // Remove the selected note from the notes lists
-            this.noteViewModel.Notes.Remove(noteToDelete);
-            this.noteViewModel.NotesForLV.Remove(noteToDelete);
-            this.noteViewModel.FirePropertyChanged("NotesForLV");
+            if (result == ContentDialogResult.Primary)
+            {
+                // Get the selected note
+                Models.NoteModel noteToDelete = this.noteViewModel.SelectedNote;
 
-            // Set the selected note to null
-            this.noteViewModel.SelectedNote = null;
+                // Remove the selected note from the notes lists
+                this.noteViewModel.Notes.Remove(noteToDelete);
+                this.noteViewModel.NotesForLV.Remove(noteToDelete);
+                this.noteViewModel.FirePropertyChanged("NotesForLV");
 
-            // Reset the notes properties
-            this.noteViewModel.NoteTitle = "";
-            this.noteViewModel.FirePropertyChanged("NoteTitle");
-            this.noteViewModel.NoteContent = "";
-            this.noteViewModel.FirePropertyChanged("NoteContent");
+                // Set the selected note to null
+                this.noteViewModel.SelectedNote = null;
 
-            // Notify that the selected note has changed
-            this.noteViewModel.FirePropertyChanged("SelectedNote");
+                // Reset the notes properties
+                this.noteViewModel.NoteTitle = "";
+                this.noteViewModel.FirePropertyChanged("NoteTitle");
+                this.noteViewModel.NoteContent = "";
+                this.noteViewModel.FirePropertyChanged("NoteContent");
 
-            // Delete the data file in the repo
-            Repositories.NotesRepo.DeleteNoteFile(noteToDelete);
+                // Notify that the selected note has changed
+                this.noteViewModel.FirePropertyChanged("SelectedNote");
+
+                // Delete the data file in the repo
+                Repositories.NotesRepo.DeleteNoteFile(noteToDelete);
+            }
         }
 
         /// <summary>
